move into qsharp package
