--- circuit_vis/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@microsoft/quantum-viz.js",
-  "version": "1.0.6",
-  "description": "quantum-viz.js is a configurable tool for visualizing quantum circuits.",
-  "main": "dist/qviz.min.js",
-  "scripts": {
-    "build": "tsc",
-    "build:prod": "tsc && webpack",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "prepublishOnly": "npm run lint && npm run test && npm run build:prod",
-    "start": "tsc -w"
-  },
-  "files": [
-    "README.md",
-    "LICENSE.txt",
-    "lib",
-    "dist"
-  ],
-  "author": "Microsoft Quantum",
-  "license": "MIT",
-  "homepage": "https://github.com/microsoft/quantum-viz.js#readme"
-}
