--- circuit_vis/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "lib": ["DOM", "ES2019"],
-    "declaration": true,
-    "strict": true,
-    "outDir": "lib",
-    "skipLibCheck": true
-  },
-  "include": ["src/**/*"]
-}
