+++ npm/qsharp/src/browser.ts
@@ -191,3 +191,5 @@ export type {
 };
 
 export * as utils from "./utils.js";
+
+export type { Circuit as CircuitData } from "./shared/circuit.js";
