+++ npm/qsharp/src/debug-service/debug-service.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { type Circuit as CircuitData } from "@microsoft/quantum-viz.js/lib/circuit.js";
+import { type Circuit as CircuitData } from "../shared/circuit.js";
 import type {
   DebugService,
   IBreakpointSpan,
