+++ npm/qsharp/src/shared/README.md
@@ -1,0 +1,1 @@
+This directory contains shared modules to be referenced from both `src/` and `ux/`.
