+++ npm/qsharp/src/shared/register.ts
