+++ npm/qsharp/ux/circuit-vis/README.md
@@ -1,0 +1,1 @@
+The code in this folder was copied and adapted from `https://github.com/microsoft/quantum-viz.js`.
