+++ npm/qsharp/ux/circuit-vis/circuit.ts
@@ -1,0 +1,7 @@
+export {
+  ConditionalRender,
+  type Circuit,
+  type DataAttributes,
+  type Operation,
+  type Qubit,
+} from "../../src/shared/circuit";
