+++ npm/qsharp/ux/circuit-vis/constants.ts
