+++ npm/qsharp/ux/circuit-vis/formatters/formatUtils.ts
