+++ npm/qsharp/ux/circuit-vis/formatters/gateFormatter.ts
