+++ npm/qsharp/ux/circuit-vis/formatters/registerFormatter.ts
