+++ npm/qsharp/ux/circuit-vis/metadata.ts
