+++ npm/qsharp/ux/circuit-vis/register.ts
@@ -1,0 +1,6 @@
+export {
+  RegisterType,
+  type Register,
+  type RegisterMap,
+  type RegisterMetadata,
+} from "../../src/shared/register";
