+++ npm/qsharp/ux/circuit-vis/sqore.ts
