+++ npm/qsharp/ux/circuit-vis/utils.ts
